introduce parser internal datamodel, separate sv specific methods
